Modularise the signup page
